refactor: moves types to a separate file
